hit reaction
you guys will need to add a left and right hit animation, just copy and
past where I made changes in mine and change the animation names
accordingly.  I changed a few things in the update, the most important
one is where the controlled part encapsulates, it is only within the
part where the actions and movements are being set.